requirements.txt
Flask
opencv-python
torch
numpy
cloudinary
flask-socketio
pymongo
gunicorn==21.2.0
flask_cors
# Pillow
pandas
requests


gitpython>=3.1.30
matplotlib>=3.3
pillow>=10.3.0
psutil
PyYAML>=5.3.1
scipy>=1.4.1
thop>=0.1.1
torchvision>=0.9.0
tqdm>=4.64.0
ultralytics>=8.0.232